% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{extract_gpx_route}
\alias{extract_gpx_route}
\title{Extract Time, Coordinates And Elevation From a GPX File}
\usage{
extract_gpx_route(gpx_file)
}
\arguments{
\item{gpx_file}{Character. Path to a valid .gpx file.}
}
\value{
sf-class data.frame with columns time, lon, lat, geometry, distance
}
\description{
Takes a .gpx file as input and extract the date, time, latitude, longitude
and elevation data to a data.frame; converts it to sf-class with the
coordinates coerced to a geometry column. Designed for use with .gpx files
downloaded from the Apple Health app, which represent individual 'workouts'.
}
\examples{
\dontrun{
x <- "apple_health_export/workout-routes/route_2021-12-25_9.31am.gpx"
extract_gpx_route(x)
}

}
