% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gpx_segment}
\alias{gpx_segment}
\title{A Workout Route Segment}
\format{
An sf-class data.frame with 501 features and 5 fields:
\describe{
\item{time}{datetime}
\item{ele}{elevation (metres)}
\item{lon}{longitude}
\item{lat}{latitude}
\item{geomatry}{sf-class POINT geometry of lon-lat}
\item{distance}{units (metres), distance between this point and the previous}
}
}
\usage{
gpx_segment
}
\description{
An sf-class data.frame with a row per point recorded along a workout route,
originally exported as a gpx file from Apple Health and with data exported
by the \code{\link{extract_gpx3d}} function.
}
\keyword{datasets}
