% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{plot_gpx3d}
\alias{plot_gpx3d}
\title{Render A 3D Plot Of A Route From A GPX File}
\usage{
plot_gpx3d(route_sf, route_only = FALSE)
}
\arguments{
\item{route_sf}{sf-class data.frame, output from \code{\link{extract_gpx3d}}.}

\item{route_only}{Logical. Retain all chart elements if \code{FALSE} (default) or
retain only the route path if \code{FALSE}.}
}
\value{
A 3D rendering of the route path in a {devoutrgl} device.
}
\description{
Create a {ggplot2} plot object with a third dimension thanks to {ggrgl}. The x and y
coordinates are the longitude and latitude, the z dimension is the elevation along
the route. The chart title includes the total distance, elevation disparity, plus the
date and start/end times.
}
\examples{
\dontrun{
x <- "~/Downloads/apple_health_export/workout-routes/route_2021-12-25_9.31am.gpx"
y <- extract_gpx3d(x)
plot_gpx3d(y)
}
}
